requirements.txt
pandas>=1.0.0
tensorflow>=2.0.0
numpy>=1.18.0
opencv-python>=4.2.0.32
Pillow>=7.0.0
matplotlib>=3.0.0
streamlit
